requirements.txt
python-nmap==0.6.4
tabulate==0.8.9
